requirements.txt
django-bootstrap-form==3.4
html5lib==1.0.1
Pillow==5.3.0
psycopg2==2.7.6.1
PyPDF2==1.26.0
pytz==2018.7
reportlab==3.5.13
six==1.12.0
virtualenv==16.1.0
webencodings==0.5.1
xhtml2pdf==0.2.3
xlwt==1.3.0
